Refactor logo component and add usePagefind hook
Resize and reposition site logo and refactor SVG into a wrapper for styling/interaction. Header: remove unused compact import, reduce overlay logo width and add top positioning for the inline logo. LogoMemorial: wrap the SVG in a div that applies width/height, className and onClick so the logo can be styled and receive events. Also add a new usePagefind hook and update the Search page to integrate the new search hook.
