+++ src/components/LogoMemorial.tsx
@@ -2,155 +2,157 @@ const LogoMemorial = ({ width = 200, color = '#64c29c', onClick, className }) =>
   const ratio = 633.89 / 624.38
   const height = width / ratio
   return (
-    <svg
-      id="Layer_4"
-      data-name="Layer 4"
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      viewBox="0 0 633.89 624.38"
-    >
-      <g>
-        <rect x="13.07" y="23.11" width="490.44" height="79.75" style={{ fill: 'white' }} />
+    <div style={{ width, height }} className={className} onClick={onClick}>
+      <svg
+        id="Layer_4"
+        data-name="Layer 4"
+        xmlns="http://www.w3.org/2000/svg"
+        width={width}
+        height={height}
+        viewBox="0 0 633.89 624.38"
+      >
+        <g>
+          <rect x="13.07" y="23.11" width="490.44" height="79.75" style={{ fill: 'white' }} />
 
-        <path
-          d="M310.65,62.23c-1.11-.66-2.39-.99-3.85-.99s-2.73.33-3.83.99c-1.09.66-1.96,1.57-2.6,2.73-.64,1.16-.96,2.54-.96,4.14s.32,2.93.96,4.11c.64,1.18,1.51,2.1,2.6,2.76,1.09.66,2.37.99,3.83.99s2.74-.33,3.85-.99c1.11-.66,1.98-1.58,2.6-2.76s.94-2.55.94-4.11-.31-2.98-.94-4.14c-.62-1.16-1.49-2.07-2.6-2.73Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M235.41,62.33c-1.13-1.07-2.65-1.61-4.56-1.61-1.49,0-2.75.31-3.77.94-1.02.62-1.8,1.56-2.32,2.81-.27.64-.46,1.35-.59,2.14h12.93c0-1.77-.56-3.2-1.69-4.27Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M365.52,71.89c-.83.64-1.25,1.48-1.25,2.52,0,1.14.38,2.01,1.15,2.6.76.59,1.84.88,3.23.88,1.98,0,3.55-.55,4.71-1.66,1.16-1.11,1.74-2.65,1.74-4.63v-1.35l-5.36.42c-1.98.17-3.38.58-4.22,1.22Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M13.07,23.11v79.75h490.44V23.11H13.07ZM232.52,45.52h5.88l-5.1,7.71h-4.68l3.9-7.71ZM212.01,82.01h-6.46v-16.53c0-1.3.02-2.4.05-3.3.03-.9.07-1.7.1-2.39.03-.7.11-1.39.21-2.08l-9.68,24.31h-6.14l-9.63-24.26c.17,1.15.29,2.6.34,4.38.05,1.77.08,3.85.08,6.25v13.64h-6.46v-38.06h6.46l12.28,30.3,12.29-30.3h6.56v38.06ZM239.36,80.29c-2.2,1.6-4.99,2.39-8.36,2.39-2.53,0-4.78-.58-6.74-1.74-1.96-1.16-3.5-2.76-4.61-4.79-1.11-2.03-1.66-4.36-1.66-7s.54-5.03,1.61-7.08c1.07-2.04,2.58-3.65,4.53-4.82,1.94-1.16,4.17-1.74,6.66-1.74s4.84.55,6.72,1.64c1.87,1.09,3.32,2.61,4.35,4.55,1.02,1.94,1.54,4.27,1.54,6.98v1.92l-19.34.05c.19,1.99.76,3.56,1.71,4.69,1.2,1.42,2.98,2.14,5.34,2.14,1.74,0,3.17-.33,4.3-.99,1.13-.66,1.86-1.59,2.21-2.81h5.88c-.56,2.81-1.94,5.01-4.14,6.61ZM288.32,82.01h-6.3v-15.15c0-1.84-.46-3.24-1.38-4.19-.92-.96-2.14-1.43-3.67-1.43-.97,0-1.87.23-2.68.68-.82.45-1.47,1.13-1.95,2.03-.49.9-.73,2.07-.73,3.49v14.58h-6.35v-15.2c0-1.8-.45-3.18-1.35-4.11-.9-.94-2.1-1.41-3.59-1.41-1.01,0-1.93.22-2.76.65-.83.44-1.49,1.11-1.98,2.03-.49.92-.73,2.07-.73,3.46v14.58h-6.35v-25.72h5.83l.52,3.05c.55-.88,1.35-1.68,2.39-2.4,1.39-.96,3.12-1.43,5.21-1.43,2.26,0,4.11.59,5.54,1.77,1.04.85,1.84,1.91,2.42,3.16.52-1.25,1.33-2.31,2.47-3.16,1.58-1.18,3.5-1.77,5.75-1.77,3.12,0,5.52.88,7.18,2.63,1.67,1.75,2.5,4.1,2.5,7.05v16.82ZM318.77,76.15c-1.18,2.03-2.8,3.62-4.87,4.76s-4.43,1.72-7.11,1.72-5-.57-7.08-1.72-3.71-2.73-4.89-4.76c-1.18-2.03-1.77-4.38-1.77-7.05s.59-5.03,1.77-7.06c1.18-2.03,2.81-3.62,4.89-4.76,2.08-1.14,4.44-1.72,7.08-1.72s5.04.57,7.11,1.72c2.06,1.15,3.69,2.73,4.87,4.76,1.18,2.03,1.77,4.38,1.77,7.06s-.59,5.02-1.77,7.05ZM341.73,62.07h-2.34c-2.32,0-4.15.62-5.47,1.85-1.32,1.23-1.98,3.1-1.98,5.6v12.5h-6.35v-25.67h5.99l.37,3.82c.44-1.04,1.08-1.93,1.97-2.67,1.28-1.07,2.95-1.62,5-1.62.45,0,.9.02,1.35.05.45.03.94.12,1.46.26v5.88ZM352.4,82.01h-6.35v-25.72h6.35v25.72ZM351.91,50.02c-.75.74-1.66,1.12-2.73,1.12s-1.99-.37-2.73-1.12c-.75-.75-1.12-1.66-1.12-2.73s.37-1.98,1.12-2.71c.75-.73,1.66-1.09,2.73-1.09s1.99.37,2.73,1.09c.75.73,1.12,1.63,1.12,2.71s-.37,1.99-1.12,2.73ZM381.24,82.01h-5.57l-.47-3.8c-.56,1.32-1.62,2.4-3.2,3.23-1.58.84-3.38,1.25-5.39,1.25-2.71,0-4.85-.74-6.43-2.21s-2.37-3.41-2.37-5.8.82-4.24,2.47-5.65c1.65-1.41,4-2.23,7.06-2.47l7.7-.62v-.57c0-1.18-.22-2.12-.65-2.84-.44-.71-1.05-1.23-1.85-1.56-.8-.33-1.74-.49-2.81-.49-1.87,0-3.31.37-4.32,1.12-1.01.75-1.51,1.8-1.51,3.15h-5.41c0-1.87.48-3.49,1.43-4.87.95-1.37,2.31-2.44,4.06-3.2,1.75-.76,3.77-1.14,6.06-1.14s4.32.41,5.99,1.22c1.66.82,2.95,2.03,3.85,3.64.9,1.62,1.35,3.62,1.35,6.01v15.62ZM393.89,82.01h-6.3v-38.73h6.3v38.73Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-      <g>
-        <path
-          d="M250.59,143.85c-2.24-1.23-4.9-1.85-7.99-1.85h-7.34v30.4h7.44c3.05,0,5.69-.62,7.91-1.87,2.22-1.25,3.95-3.01,5.18-5.28,1.23-2.27,1.85-4.96,1.85-8.04s-.62-5.78-1.85-8.07c-1.23-2.29-2.97-4.05-5.2-5.28Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M339.07,165.37c-1.04.97-1.56,2.15-1.56,3.54,0,1.53.52,2.68,1.56,3.46,1.04.78,2.43,1.17,4.17,1.17,2.67,0,4.74-.85,6.19-2.55,1.46-1.7,2.19-3.9,2.19-6.61v-1.35l-7.44.57c-2.36.21-4.06.8-5.1,1.77Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M184.86,122.79v79.75h435.95v-79.75H184.86ZM259.18,166.96c-1.54,2.79-3.71,4.97-6.51,6.53-2.79,1.56-6.06,2.34-9.81,2.34h-11.35v-37.28h11.25c3.75,0,7.03.78,9.84,2.35,2.81,1.56,5,3.74,6.56,6.53,1.56,2.79,2.34,6.05,2.34,9.76s-.77,6.96-2.32,9.76ZM271.08,175.84h-3.59v-25.2h3.59v25.2ZM271.15,142.94c-.54.52-1.17.78-1.9.78s-1.35-.26-1.87-.78-.78-1.15-.78-1.87.26-1.36.78-1.9c.52-.54,1.14-.81,1.87-.81s1.36.27,1.9.81c.54.54.81,1.17.81,1.9s-.27,1.35-.81,1.87ZM301.48,175.06c0,2.5-.5,4.68-1.51,6.53-1.01,1.86-2.42,3.3-4.24,4.32-1.82,1.02-3.97,1.54-6.43,1.54-3.23,0-5.9-.87-8.02-2.6-2.12-1.74-3.44-4.11-3.95-7.13h3.59c.38,2.05,1.31,3.63,2.79,4.76,1.47,1.13,3.36,1.69,5.65,1.69,2.6,0,4.69-.8,6.25-2.4,1.56-1.59,2.34-3.71,2.34-6.35v-6.02c-.75,1.55-1.83,2.83-3.25,3.81-1.79,1.23-3.88,1.85-6.27,1.85s-4.46-.57-6.19-1.72c-1.74-1.15-3.06-2.67-3.98-4.56-.92-1.89-1.38-3.98-1.38-6.27s.47-4.38,1.41-6.27c.94-1.89,2.29-3.41,4.06-4.55,1.77-1.14,3.89-1.72,6.35-1.72s4.44.59,6.14,1.77c1.4.97,2.46,2.32,3.2,4l.34-5.1h3.12v24.42ZM312.57,175.84h-3.59v-25.2h3.59v25.2ZM312.65,142.94c-.54.52-1.17.78-1.9.78s-1.35-.26-1.87-.78c-.52-.52-.78-1.15-.78-1.87s.26-1.36.78-1.9c.52-.54,1.14-.81,1.87-.81s1.36.27,1.9.81c.54.54.81,1.17.81,1.9s-.27,1.35-.81,1.87ZM330.89,153.66h-5.05v22.18h-3.54v-22.18h-5.05v-3.02h5.05v-7.96h3.54v7.96h5.05v3.02ZM355.15,175.84h-3.02l-.47-4.43c-.76,1.6-1.97,2.84-3.62,3.72-1.65.89-3.53,1.33-5.65,1.33-2.67,0-4.76-.69-6.27-2.08-1.51-1.39-2.27-3.18-2.27-5.36,0-2.32.83-4.18,2.5-5.57s3.97-2.2,6.92-2.45l8.33-.68v-.84c0-1.66-.3-2.97-.89-3.93-.59-.96-1.39-1.62-2.4-2-1.01-.38-2.13-.57-3.38-.57-2.22,0-3.94.47-5.16,1.41-1.21.94-1.82,2.26-1.82,3.95h-3.28c0-1.7.44-3.17,1.3-4.42.87-1.25,2.08-2.22,3.64-2.92,1.56-.69,3.37-1.04,5.41-1.04s3.71.34,5.21,1.02c1.49.68,2.68,1.72,3.57,3.12.89,1.41,1.33,3.22,1.33,5.44v16.29ZM366.08,175.84h-3.59v-37.95h3.59v37.95Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M293.7,154.36c-1.27-.81-2.84-1.22-4.71-1.22s-3.34.42-4.61,1.27c-1.27.85-2.24,1.98-2.91,3.38-.68,1.41-1.02,2.96-1.02,4.66s.34,3.31,1.02,4.74c.68,1.42,1.64,2.55,2.91,3.38,1.27.83,2.79,1.25,4.56,1.25s3.36-.41,4.66-1.22c1.3-.81,2.3-1.92,2.99-3.33.69-1.41,1.04-3.01,1.04-4.82s-.34-3.4-1.02-4.79c-.68-1.39-1.65-2.49-2.91-3.31Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-      <g>
-        <path
-          d="M344.45,553.22c-1.37-.87-2.96-1.3-4.76-1.3s-3.39.43-4.76,1.3c-1.37.87-2.45,2.05-3.23,3.54-.78,1.49-1.17,3.21-1.17,5.15s.39,3.72,1.17,5.23c.78,1.51,1.86,2.69,3.23,3.54,1.37.85,2.96,1.27,4.76,1.27s3.39-.42,4.76-1.27c1.37-.85,2.45-2.03,3.23-3.54.78-1.51,1.17-3.25,1.17-5.23s-.39-3.66-1.17-5.15c-.78-1.49-1.86-2.67-3.23-3.54Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M242.94,553.97c-1.44-1.47-3.34-2.21-5.7-2.21-1.74,0-3.25.4-4.55,1.2-1.3.8-2.3,1.93-2.99,3.41-.48,1.02-.78,2.18-.93,3.46h16.34c0-2.43-.72-4.38-2.16-5.86Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M315.33,553.22c-1.29-.87-2.85-1.3-4.69-1.3s-3.41.43-4.71,1.3c-1.3.87-2.3,2.05-2.99,3.54-.7,1.49-1.04,3.21-1.04,5.15s.35,3.67,1.04,5.18c.69,1.51,1.69,2.69,2.99,3.54,1.3.85,2.87,1.27,4.71,1.27s3.4-.42,4.69-1.27c1.28-.85,2.26-2.03,2.94-3.54.68-1.51,1.01-3.24,1.01-5.18s-.34-3.66-1.01-5.15c-.68-1.49-1.66-2.67-2.94-3.54Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M83.96,521.52v79.75h435.95v-79.75H83.96ZM170.5,574.56h-20.88v-37.28h3.75v33.84h17.13v3.44ZM196.64,574.56h-3.12l-.47-4.27c-.7,1.46-1.85,2.64-3.46,3.54-1.61.9-3.41,1.35-5.39,1.35-2.95,0-5.22-.94-6.79-2.81-1.58-1.87-2.37-4.34-2.37-7.39v-15.67h3.64v14.53c0,2.98.62,5.08,1.85,6.3,1.23,1.22,2.84,1.82,4.82,1.82,2.43,0,4.32-.79,5.68-2.37,1.35-1.58,2.03-3.86,2.03-6.84v-13.38h3.59v25.2ZM223.5,574.56h-4.27l-7.08-10.1-7.03,10.1h-4.22l9.01-12.6-8.95-12.6h4.22l7.13,10.1,7.08-10.1h4.17l-8.95,12.65,8.9,12.55ZM248.69,562.44h-20.02c.12,2.93.88,5.24,2.29,6.92,1.55,1.84,3.74,2.76,6.59,2.76,2.05,0,3.73-.43,5.05-1.3,1.32-.87,2.2-2.14,2.65-3.8h3.44c-.66,2.64-1.94,4.66-3.85,6.06-1.91,1.41-4.36,2.11-7.34,2.11-2.43,0-4.57-.56-6.43-1.67-1.86-1.11-3.31-2.65-4.35-4.63-1.04-1.98-1.56-4.28-1.56-6.92s.51-4.9,1.53-6.9c1.02-1.99,2.44-3.56,4.24-4.68s3.91-1.69,6.3-1.69,4.3.5,6.01,1.51c1.72,1.01,3.05,2.4,4.01,4.19.95,1.79,1.43,3.88,1.43,6.27v1.77ZM291.02,574.56h-3.49v-15.77c0-2.15-.56-3.83-1.67-5.05-1.11-1.22-2.64-1.82-4.58-1.82-1.39,0-2.59.32-3.59.96-1.01.64-1.78,1.5-2.32,2.58-.54,1.07-.81,2.26-.81,3.54v15.56h-3.54v-15.83c0-2.15-.56-3.82-1.69-5-1.13-1.18-2.65-1.77-4.56-1.77-1.39,0-2.59.31-3.62.94-1.02.62-1.8,1.47-2.34,2.53-.54,1.06-.81,2.23-.81,3.51v15.62h-3.59v-25.2h3.12l.4,3.63c.56-1.16,1.4-2.12,2.54-2.88,1.44-.96,3.13-1.43,5.07-1.43,2.26,0,4.14.62,5.65,1.85,1.14.93,1.96,2.1,2.48,3.53.48-1.43,1.31-2.6,2.47-3.53,1.54-1.23,3.48-1.85,5.8-1.85,2.71,0,4.89.83,6.56,2.5,1.66,1.67,2.5,3.96,2.5,6.87v16.5ZM321.52,568.76c-.94,1.99-2.27,3.57-4.01,4.71-1.74,1.14-3.83,1.72-6.3,1.72-2.08,0-3.96-.48-5.62-1.43-1.66-.96-2.95-2.4-3.85-4.35l-.36,5.15h-3.12v-37.95h3.59v17.65c.87-1.87,2.12-3.27,3.77-4.19,1.65-.92,3.55-1.38,5.7-1.38,2.46,0,4.56.58,6.27,1.74,1.72,1.17,3.04,2.75,3.96,4.76.92,2.01,1.38,4.28,1.38,6.82s-.47,4.75-1.41,6.74ZM350.83,568.76c-1.11,1.99-2.63,3.57-4.56,4.71s-4.12,1.72-6.59,1.72-4.62-.57-6.56-1.72c-1.94-1.14-3.47-2.72-4.58-4.71-1.11-2-1.66-4.26-1.66-6.79s.55-4.89,1.66-6.87c1.11-1.98,2.64-3.54,4.58-4.68,1.94-1.14,4.13-1.72,6.56-1.72s4.66.57,6.59,1.72,3.44,2.71,4.56,4.68c1.11,1.98,1.66,4.27,1.66,6.87s-.56,4.8-1.66,6.79ZM378.89,574.56h-3.12l-.47-4.27c-.7,1.46-1.85,2.64-3.46,3.54-1.61.9-3.41,1.35-5.39,1.35-2.95,0-5.22-.94-6.79-2.81-1.58-1.87-2.37-4.34-2.37-7.39v-15.67h3.64v14.53c0,2.98.62,5.08,1.85,6.3,1.23,1.22,2.84,1.82,4.82,1.82,2.43,0,4.32-.79,5.68-2.37,1.35-1.58,2.03-3.86,2.03-6.84v-13.38h3.59v25.2ZM399.55,552.28h-1.98c-2.33,0-4.17.71-5.54,2.14-1.37,1.42-2.06,3.35-2.06,5.78v14.37h-3.59v-25.2h3.38l.25,3.95c.44-1.18,1.18-2.18,2.25-2.99,1.35-1.02,3-1.54,4.95-1.54.38,0,.76.02,1.15.05.38.04.78.11,1.2.21v3.23ZM426.78,573.78c0,2.5-.5,4.68-1.51,6.53-1.01,1.86-2.42,3.3-4.24,4.32-1.82,1.02-3.97,1.54-6.43,1.54-3.23,0-5.9-.87-8.02-2.6-2.12-1.74-3.44-4.11-3.96-7.13h3.59c.38,2.05,1.31,3.63,2.79,4.76,1.47,1.13,3.36,1.69,5.65,1.69,2.6,0,4.69-.8,6.25-2.4,1.56-1.59,2.34-3.71,2.34-6.35v-6.02c-.75,1.55-1.83,2.82-3.25,3.81-1.79,1.23-3.88,1.85-6.27,1.85s-4.46-.57-6.19-1.72c-1.74-1.15-3.06-2.67-3.98-4.56-.92-1.89-1.38-3.98-1.38-6.27s.47-4.38,1.41-6.27c.94-1.89,2.29-3.41,4.06-4.55s3.89-1.72,6.35-1.72,4.44.59,6.14,1.77c1.4.97,2.45,2.32,3.2,4l.34-5.09h3.12v24.42Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M419,553.09c-1.27-.81-2.84-1.22-4.71-1.22s-3.34.42-4.61,1.27c-1.27.85-2.24,1.98-2.91,3.38-.68,1.41-1.01,2.96-1.01,4.66s.34,3.31,1.01,4.74c.68,1.42,1.65,2.55,2.91,3.38,1.27.83,2.79,1.25,4.56,1.25s3.36-.41,4.66-1.22c1.3-.81,2.3-1.92,2.99-3.33.69-1.41,1.04-3.01,1.04-4.82s-.34-3.4-1.01-4.79-1.65-2.49-2.91-3.31Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-      <g>
-        <path
-          d="M245.29,260.42c-1.44-1.47-3.34-2.21-5.7-2.21-1.74,0-3.25.4-4.55,1.2-1.3.8-2.3,1.93-2.99,3.41-.48,1.02-.78,2.18-.93,3.46h16.34c0-2.43-.72-4.38-2.16-5.86Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M36.37,222.47v79.75h414.32v-79.75H36.37ZM221.79,281.01h-3.12l-.36-5.15c-.9,1.94-2.19,3.39-3.85,4.35-1.66.96-3.56,1.43-5.67,1.43-2.43,0-4.51-.57-6.25-1.72-1.74-1.14-3.07-2.72-4.01-4.71-.94-2-1.41-4.24-1.41-6.74s.47-4.81,1.41-6.82,2.28-3.6,4.03-4.76c1.75-1.16,3.88-1.74,6.38-1.74,2.08,0,3.93.45,5.54,1.35,1.61.9,2.85,2.27,3.72,4.11v-17.54h3.59v37.95ZM251.05,268.88h-20.02c.12,2.93.88,5.24,2.29,6.92,1.55,1.84,3.74,2.76,6.59,2.76,2.05,0,3.73-.43,5.05-1.3,1.32-.87,2.2-2.14,2.65-3.8h3.44c-.66,2.64-1.94,4.66-3.85,6.06-1.91,1.41-4.36,2.11-7.34,2.11-2.43,0-4.57-.56-6.43-1.67-1.86-1.11-3.31-2.65-4.35-4.63-1.04-1.98-1.56-4.28-1.56-6.92s.51-4.9,1.53-6.9c1.02-1.99,2.44-3.56,4.24-4.68,1.8-1.13,3.91-1.69,6.3-1.69s4.3.5,6.01,1.51,3.05,2.4,4.01,4.19c.95,1.79,1.43,3.88,1.43,6.27v1.77Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M214.14,259.67c-1.28-.87-2.86-1.3-4.74-1.3s-3.36.43-4.66,1.3c-1.3.87-2.29,2.05-2.97,3.54-.68,1.49-1.01,3.21-1.01,5.15s.34,3.67,1.01,5.18c.68,1.51,1.66,2.69,2.97,3.54,1.3.85,2.85,1.27,4.66,1.27s3.45-.42,4.74-1.27c1.28-.85,2.27-2.03,2.97-3.54.69-1.51,1.04-3.24,1.04-5.18s-.35-3.66-1.04-5.15c-.7-1.49-1.68-2.67-2.97-3.54Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-      <g>
-        <path
-          d="M15.58,322.16v79.75h476.31v-79.75H15.58ZM236.25,380.69h-3.59v-37.95h3.59v37.95ZM263.89,380.69h-3.02l-.47-4.43c-.76,1.6-1.97,2.84-3.62,3.72-1.65.89-3.53,1.33-5.65,1.33-2.67,0-4.76-.69-6.27-2.08-1.51-1.39-2.27-3.18-2.27-5.36,0-2.32.83-4.18,2.5-5.57s3.97-2.2,6.92-2.45l8.33-.68v-.84c0-1.66-.3-2.97-.89-3.93-.59-.96-1.39-1.62-2.4-2-1.01-.38-2.13-.57-3.38-.57-2.22,0-3.94.47-5.16,1.41-1.21.94-1.82,2.26-1.82,3.95h-3.28c0-1.7.44-3.17,1.3-4.42.87-1.25,2.08-2.22,3.64-2.92,1.56-.69,3.37-1.04,5.41-1.04s3.71.34,5.21,1.02c1.49.68,2.68,1.72,3.57,3.12.89,1.41,1.33,3.22,1.33,5.44v16.29Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M247.81,370.23c-1.04.97-1.56,2.15-1.56,3.54,0,1.53.52,2.68,1.56,3.46,1.04.78,2.43,1.17,4.17,1.17,2.67,0,4.74-.85,6.19-2.55,1.46-1.7,2.19-3.9,2.19-6.61v-1.35l-7.44.57c-2.36.21-4.06.8-5.1,1.77Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-      <g>
-        <path
-          d="M184.86,421.84v79.75h265.12v-79.75H184.86ZM234.75,475.9c-1.15,1.77-2.77,3.13-4.87,4.09-2.1.96-4.56,1.43-7.37,1.43s-5.01-.46-7.03-1.38c-2.01-.92-3.58-2.21-4.71-3.88-1.13-1.66-1.73-3.62-1.8-5.88h6.67c.03,1.6.67,2.86,1.9,3.78,1.23.92,2.91,1.38,5.02,1.38,2.22,0,3.98-.46,5.28-1.38,1.3-.92,1.95-2.14,1.95-3.67,0-1.32-.35-2.37-1.04-3.15-.7-.78-1.82-1.33-3.38-1.64l-6.04-1.35c-3.23-.69-5.67-1.95-7.34-3.77-1.66-1.82-2.5-4.21-2.5-7.16,0-2.22.57-4.18,1.72-5.88,1.15-1.7,2.73-3.03,4.76-3.98,2.03-.95,4.38-1.43,7.05-1.43s4.8.45,6.69,1.35c1.89.9,3.37,2.19,4.42,3.85,1.06,1.66,1.62,3.64,1.69,5.93h-6.66c-.07-1.63-.66-2.91-1.77-3.83-1.11-.92-2.6-1.38-4.48-1.38-2.05,0-3.69.46-4.92,1.38-1.23.92-1.85,2.18-1.85,3.78,0,1.35.37,2.42,1.12,3.2.74.78,1.92,1.36,3.51,1.75l5.99,1.35c3.26.7,5.69,1.88,7.29,3.57,1.6,1.68,2.39,3.95,2.39,6.79,0,2.33-.57,4.37-1.72,6.14ZM265.99,480.74h-6.35v-14.31c0-1.46-.23-2.66-.68-3.59-.45-.94-1.08-1.64-1.87-2.11-.8-.47-1.74-.7-2.81-.7-1.32,0-2.45.29-3.41.86-.96.57-1.68,1.36-2.19,2.37-.5,1.01-.76,2.15-.76,3.44v14.06h-6.35v-38.73h6.4v16.35c.8-1.25,1.92-2.25,3.36-3,1.44-.74,3.1-1.12,4.97-1.12,3.16,0,5.56.96,7.21,2.86,1.65,1.91,2.47,4.51,2.47,7.81v15.83ZM296.44,474.88c-1.18,2.03-2.8,3.62-4.87,4.76s-4.43,1.72-7.11,1.72-5-.57-7.08-1.72c-2.08-1.14-3.71-2.73-4.89-4.76-1.18-2.03-1.77-4.38-1.77-7.05s.59-5.03,1.77-7.06c1.18-2.03,2.81-3.62,4.89-4.76,2.08-1.14,4.44-1.72,7.08-1.72s5.04.57,7.11,1.72c2.06,1.15,3.69,2.73,4.87,4.76,1.18,2.03,1.77,4.38,1.77,7.06s-.59,5.02-1.77,7.05ZM325.54,480.74h-5.57l-.47-3.8c-.56,1.32-1.62,2.4-3.2,3.23-1.58.84-3.38,1.25-5.39,1.25-2.71,0-4.85-.74-6.43-2.21-1.58-1.47-2.37-3.41-2.37-5.8s.82-4.24,2.47-5.65c1.65-1.41,4-2.23,7.06-2.47l7.7-.62v-.57c0-1.18-.22-2.12-.65-2.84-.44-.71-1.05-1.23-1.85-1.56-.8-.33-1.74-.49-2.81-.49-1.87,0-3.31.37-4.32,1.12-1.01.75-1.51,1.8-1.51,3.15h-5.41c0-1.87.48-3.49,1.43-4.87.95-1.37,2.31-2.44,4.06-3.2,1.75-.76,3.77-1.14,6.06-1.14s4.32.41,5.99,1.22c1.66.82,2.95,2.03,3.85,3.64.9,1.62,1.35,3.62,1.35,6.01v15.62ZM356.2,480.74h-6.35v-14.31c0-1.46-.23-2.66-.68-3.59-.45-.94-1.08-1.64-1.87-2.11-.8-.47-1.74-.7-2.81-.7-1.32,0-2.45.29-3.41.86-.96.57-1.68,1.36-2.19,2.37s-.76,2.15-.76,3.44v14.06h-6.35v-38.73h6.4v16.35c.8-1.25,1.92-2.25,3.36-3,1.44-.74,3.1-1.12,4.97-1.12,3.16,0,5.56.96,7.21,2.86,1.65,1.91,2.47,4.51,2.47,7.81v15.83Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M309.82,470.62c-.83.64-1.25,1.48-1.25,2.52,0,1.14.38,2.01,1.15,2.6.76.59,1.84.88,3.23.88,1.98,0,3.55-.55,4.71-1.66,1.16-1.11,1.74-2.65,1.74-4.63v-1.35l-5.36.42c-1.98.17-3.38.58-4.22,1.22Z"
-          style={{
-            fill: color,
-          }}
-        />
-        <path
-          d="M288.32,460.96c-1.11-.66-2.39-.99-3.85-.99s-2.73.33-3.83.99c-1.09.66-1.96,1.57-2.6,2.73-.64,1.16-.96,2.54-.96,4.14s.32,2.93.96,4.11,1.51,2.1,2.6,2.76c1.09.66,2.37.99,3.83.99s2.74-.33,3.85-.99c1.11-.66,1.98-1.58,2.6-2.76.62-1.18.94-2.55.94-4.11s-.31-2.98-.94-4.14c-.62-1.16-1.49-2.07-2.6-2.73Z"
-          style={{
-            fill: color,
-          }}
-        />
-      </g>
-    </svg>
+          <path
+            d="M310.65,62.23c-1.11-.66-2.39-.99-3.85-.99s-2.73.33-3.83.99c-1.09.66-1.96,1.57-2.6,2.73-.64,1.16-.96,2.54-.96,4.14s.32,2.93.96,4.11c.64,1.18,1.51,2.1,2.6,2.76,1.09.66,2.37.99,3.83.99s2.74-.33,3.85-.99c1.11-.66,1.98-1.58,2.6-2.76s.94-2.55.94-4.11-.31-2.98-.94-4.14c-.62-1.16-1.49-2.07-2.6-2.73Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M235.41,62.33c-1.13-1.07-2.65-1.61-4.56-1.61-1.49,0-2.75.31-3.77.94-1.02.62-1.8,1.56-2.32,2.81-.27.64-.46,1.35-.59,2.14h12.93c0-1.77-.56-3.2-1.69-4.27Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M365.52,71.89c-.83.64-1.25,1.48-1.25,2.52,0,1.14.38,2.01,1.15,2.6.76.59,1.84.88,3.23.88,1.98,0,3.55-.55,4.71-1.66,1.16-1.11,1.74-2.65,1.74-4.63v-1.35l-5.36.42c-1.98.17-3.38.58-4.22,1.22Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M13.07,23.11v79.75h490.44V23.11H13.07ZM232.52,45.52h5.88l-5.1,7.71h-4.68l3.9-7.71ZM212.01,82.01h-6.46v-16.53c0-1.3.02-2.4.05-3.3.03-.9.07-1.7.1-2.39.03-.7.11-1.39.21-2.08l-9.68,24.31h-6.14l-9.63-24.26c.17,1.15.29,2.6.34,4.38.05,1.77.08,3.85.08,6.25v13.64h-6.46v-38.06h6.46l12.28,30.3,12.29-30.3h6.56v38.06ZM239.36,80.29c-2.2,1.6-4.99,2.39-8.36,2.39-2.53,0-4.78-.58-6.74-1.74-1.96-1.16-3.5-2.76-4.61-4.79-1.11-2.03-1.66-4.36-1.66-7s.54-5.03,1.61-7.08c1.07-2.04,2.58-3.65,4.53-4.82,1.94-1.16,4.17-1.74,6.66-1.74s4.84.55,6.72,1.64c1.87,1.09,3.32,2.61,4.35,4.55,1.02,1.94,1.54,4.27,1.54,6.98v1.92l-19.34.05c.19,1.99.76,3.56,1.71,4.69,1.2,1.42,2.98,2.14,5.34,2.14,1.74,0,3.17-.33,4.3-.99,1.13-.66,1.86-1.59,2.21-2.81h5.88c-.56,2.81-1.94,5.01-4.14,6.61ZM288.32,82.01h-6.3v-15.15c0-1.84-.46-3.24-1.38-4.19-.92-.96-2.14-1.43-3.67-1.43-.97,0-1.87.23-2.68.68-.82.45-1.47,1.13-1.95,2.03-.49.9-.73,2.07-.73,3.49v14.58h-6.35v-15.2c0-1.8-.45-3.18-1.35-4.11-.9-.94-2.1-1.41-3.59-1.41-1.01,0-1.93.22-2.76.65-.83.44-1.49,1.11-1.98,2.03-.49.92-.73,2.07-.73,3.46v14.58h-6.35v-25.72h5.83l.52,3.05c.55-.88,1.35-1.68,2.39-2.4,1.39-.96,3.12-1.43,5.21-1.43,2.26,0,4.11.59,5.54,1.77,1.04.85,1.84,1.91,2.42,3.16.52-1.25,1.33-2.31,2.47-3.16,1.58-1.18,3.5-1.77,5.75-1.77,3.12,0,5.52.88,7.18,2.63,1.67,1.75,2.5,4.1,2.5,7.05v16.82ZM318.77,76.15c-1.18,2.03-2.8,3.62-4.87,4.76s-4.43,1.72-7.11,1.72-5-.57-7.08-1.72-3.71-2.73-4.89-4.76c-1.18-2.03-1.77-4.38-1.77-7.05s.59-5.03,1.77-7.06c1.18-2.03,2.81-3.62,4.89-4.76,2.08-1.14,4.44-1.72,7.08-1.72s5.04.57,7.11,1.72c2.06,1.15,3.69,2.73,4.87,4.76,1.18,2.03,1.77,4.38,1.77,7.06s-.59,5.02-1.77,7.05ZM341.73,62.07h-2.34c-2.32,0-4.15.62-5.47,1.85-1.32,1.23-1.98,3.1-1.98,5.6v12.5h-6.35v-25.67h5.99l.37,3.82c.44-1.04,1.08-1.93,1.97-2.67,1.28-1.07,2.95-1.62,5-1.62.45,0,.9.02,1.35.05.45.03.94.12,1.46.26v5.88ZM352.4,82.01h-6.35v-25.72h6.35v25.72ZM351.91,50.02c-.75.74-1.66,1.12-2.73,1.12s-1.99-.37-2.73-1.12c-.75-.75-1.12-1.66-1.12-2.73s.37-1.98,1.12-2.71c.75-.73,1.66-1.09,2.73-1.09s1.99.37,2.73,1.09c.75.73,1.12,1.63,1.12,2.71s-.37,1.99-1.12,2.73ZM381.24,82.01h-5.57l-.47-3.8c-.56,1.32-1.62,2.4-3.2,3.23-1.58.84-3.38,1.25-5.39,1.25-2.71,0-4.85-.74-6.43-2.21s-2.37-3.41-2.37-5.8.82-4.24,2.47-5.65c1.65-1.41,4-2.23,7.06-2.47l7.7-.62v-.57c0-1.18-.22-2.12-.65-2.84-.44-.71-1.05-1.23-1.85-1.56-.8-.33-1.74-.49-2.81-.49-1.87,0-3.31.37-4.32,1.12-1.01.75-1.51,1.8-1.51,3.15h-5.41c0-1.87.48-3.49,1.43-4.87.95-1.37,2.31-2.44,4.06-3.2,1.75-.76,3.77-1.14,6.06-1.14s4.32.41,5.99,1.22c1.66.82,2.95,2.03,3.85,3.64.9,1.62,1.35,3.62,1.35,6.01v15.62ZM393.89,82.01h-6.3v-38.73h6.3v38.73Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+        <g>
+          <path
+            d="M250.59,143.85c-2.24-1.23-4.9-1.85-7.99-1.85h-7.34v30.4h7.44c3.05,0,5.69-.62,7.91-1.87,2.22-1.25,3.95-3.01,5.18-5.28,1.23-2.27,1.85-4.96,1.85-8.04s-.62-5.78-1.85-8.07c-1.23-2.29-2.97-4.05-5.2-5.28Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M339.07,165.37c-1.04.97-1.56,2.15-1.56,3.54,0,1.53.52,2.68,1.56,3.46,1.04.78,2.43,1.17,4.17,1.17,2.67,0,4.74-.85,6.19-2.55,1.46-1.7,2.19-3.9,2.19-6.61v-1.35l-7.44.57c-2.36.21-4.06.8-5.1,1.77Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M184.86,122.79v79.75h435.95v-79.75H184.86ZM259.18,166.96c-1.54,2.79-3.71,4.97-6.51,6.53-2.79,1.56-6.06,2.34-9.81,2.34h-11.35v-37.28h11.25c3.75,0,7.03.78,9.84,2.35,2.81,1.56,5,3.74,6.56,6.53,1.56,2.79,2.34,6.05,2.34,9.76s-.77,6.96-2.32,9.76ZM271.08,175.84h-3.59v-25.2h3.59v25.2ZM271.15,142.94c-.54.52-1.17.78-1.9.78s-1.35-.26-1.87-.78-.78-1.15-.78-1.87.26-1.36.78-1.9c.52-.54,1.14-.81,1.87-.81s1.36.27,1.9.81c.54.54.81,1.17.81,1.9s-.27,1.35-.81,1.87ZM301.48,175.06c0,2.5-.5,4.68-1.51,6.53-1.01,1.86-2.42,3.3-4.24,4.32-1.82,1.02-3.97,1.54-6.43,1.54-3.23,0-5.9-.87-8.02-2.6-2.12-1.74-3.44-4.11-3.95-7.13h3.59c.38,2.05,1.31,3.63,2.79,4.76,1.47,1.13,3.36,1.69,5.65,1.69,2.6,0,4.69-.8,6.25-2.4,1.56-1.59,2.34-3.71,2.34-6.35v-6.02c-.75,1.55-1.83,2.83-3.25,3.81-1.79,1.23-3.88,1.85-6.27,1.85s-4.46-.57-6.19-1.72c-1.74-1.15-3.06-2.67-3.98-4.56-.92-1.89-1.38-3.98-1.38-6.27s.47-4.38,1.41-6.27c.94-1.89,2.29-3.41,4.06-4.55,1.77-1.14,3.89-1.72,6.35-1.72s4.44.59,6.14,1.77c1.4.97,2.46,2.32,3.2,4l.34-5.1h3.12v24.42ZM312.57,175.84h-3.59v-25.2h3.59v25.2ZM312.65,142.94c-.54.52-1.17.78-1.9.78s-1.35-.26-1.87-.78c-.52-.52-.78-1.15-.78-1.87s.26-1.36.78-1.9c.52-.54,1.14-.81,1.87-.81s1.36.27,1.9.81c.54.54.81,1.17.81,1.9s-.27,1.35-.81,1.87ZM330.89,153.66h-5.05v22.18h-3.54v-22.18h-5.05v-3.02h5.05v-7.96h3.54v7.96h5.05v3.02ZM355.15,175.84h-3.02l-.47-4.43c-.76,1.6-1.97,2.84-3.62,3.72-1.65.89-3.53,1.33-5.65,1.33-2.67,0-4.76-.69-6.27-2.08-1.51-1.39-2.27-3.18-2.27-5.36,0-2.32.83-4.18,2.5-5.57s3.97-2.2,6.92-2.45l8.33-.68v-.84c0-1.66-.3-2.97-.89-3.93-.59-.96-1.39-1.62-2.4-2-1.01-.38-2.13-.57-3.38-.57-2.22,0-3.94.47-5.16,1.41-1.21.94-1.82,2.26-1.82,3.95h-3.28c0-1.7.44-3.17,1.3-4.42.87-1.25,2.08-2.22,3.64-2.92,1.56-.69,3.37-1.04,5.41-1.04s3.71.34,5.21,1.02c1.49.68,2.68,1.72,3.57,3.12.89,1.41,1.33,3.22,1.33,5.44v16.29ZM366.08,175.84h-3.59v-37.95h3.59v37.95Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M293.7,154.36c-1.27-.81-2.84-1.22-4.71-1.22s-3.34.42-4.61,1.27c-1.27.85-2.24,1.98-2.91,3.38-.68,1.41-1.02,2.96-1.02,4.66s.34,3.31,1.02,4.74c.68,1.42,1.64,2.55,2.91,3.38,1.27.83,2.79,1.25,4.56,1.25s3.36-.41,4.66-1.22c1.3-.81,2.3-1.92,2.99-3.33.69-1.41,1.04-3.01,1.04-4.82s-.34-3.4-1.02-4.79c-.68-1.39-1.65-2.49-2.91-3.31Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+        <g>
+          <path
+            d="M344.45,553.22c-1.37-.87-2.96-1.3-4.76-1.3s-3.39.43-4.76,1.3c-1.37.87-2.45,2.05-3.23,3.54-.78,1.49-1.17,3.21-1.17,5.15s.39,3.72,1.17,5.23c.78,1.51,1.86,2.69,3.23,3.54,1.37.85,2.96,1.27,4.76,1.27s3.39-.42,4.76-1.27c1.37-.85,2.45-2.03,3.23-3.54.78-1.51,1.17-3.25,1.17-5.23s-.39-3.66-1.17-5.15c-.78-1.49-1.86-2.67-3.23-3.54Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M242.94,553.97c-1.44-1.47-3.34-2.21-5.7-2.21-1.74,0-3.25.4-4.55,1.2-1.3.8-2.3,1.93-2.99,3.41-.48,1.02-.78,2.18-.93,3.46h16.34c0-2.43-.72-4.38-2.16-5.86Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M315.33,553.22c-1.29-.87-2.85-1.3-4.69-1.3s-3.41.43-4.71,1.3c-1.3.87-2.3,2.05-2.99,3.54-.7,1.49-1.04,3.21-1.04,5.15s.35,3.67,1.04,5.18c.69,1.51,1.69,2.69,2.99,3.54,1.3.85,2.87,1.27,4.71,1.27s3.4-.42,4.69-1.27c1.28-.85,2.26-2.03,2.94-3.54.68-1.51,1.01-3.24,1.01-5.18s-.34-3.66-1.01-5.15c-.68-1.49-1.66-2.67-2.94-3.54Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M83.96,521.52v79.75h435.95v-79.75H83.96ZM170.5,574.56h-20.88v-37.28h3.75v33.84h17.13v3.44ZM196.64,574.56h-3.12l-.47-4.27c-.7,1.46-1.85,2.64-3.46,3.54-1.61.9-3.41,1.35-5.39,1.35-2.95,0-5.22-.94-6.79-2.81-1.58-1.87-2.37-4.34-2.37-7.39v-15.67h3.64v14.53c0,2.98.62,5.08,1.85,6.3,1.23,1.22,2.84,1.82,4.82,1.82,2.43,0,4.32-.79,5.68-2.37,1.35-1.58,2.03-3.86,2.03-6.84v-13.38h3.59v25.2ZM223.5,574.56h-4.27l-7.08-10.1-7.03,10.1h-4.22l9.01-12.6-8.95-12.6h4.22l7.13,10.1,7.08-10.1h4.17l-8.95,12.65,8.9,12.55ZM248.69,562.44h-20.02c.12,2.93.88,5.24,2.29,6.92,1.55,1.84,3.74,2.76,6.59,2.76,2.05,0,3.73-.43,5.05-1.3,1.32-.87,2.2-2.14,2.65-3.8h3.44c-.66,2.64-1.94,4.66-3.85,6.06-1.91,1.41-4.36,2.11-7.34,2.11-2.43,0-4.57-.56-6.43-1.67-1.86-1.11-3.31-2.65-4.35-4.63-1.04-1.98-1.56-4.28-1.56-6.92s.51-4.9,1.53-6.9c1.02-1.99,2.44-3.56,4.24-4.68s3.91-1.69,6.3-1.69,4.3.5,6.01,1.51c1.72,1.01,3.05,2.4,4.01,4.19.95,1.79,1.43,3.88,1.43,6.27v1.77ZM291.02,574.56h-3.49v-15.77c0-2.15-.56-3.83-1.67-5.05-1.11-1.22-2.64-1.82-4.58-1.82-1.39,0-2.59.32-3.59.96-1.01.64-1.78,1.5-2.32,2.58-.54,1.07-.81,2.26-.81,3.54v15.56h-3.54v-15.83c0-2.15-.56-3.82-1.69-5-1.13-1.18-2.65-1.77-4.56-1.77-1.39,0-2.59.31-3.62.94-1.02.62-1.8,1.47-2.34,2.53-.54,1.06-.81,2.23-.81,3.51v15.62h-3.59v-25.2h3.12l.4,3.63c.56-1.16,1.4-2.12,2.54-2.88,1.44-.96,3.13-1.43,5.07-1.43,2.26,0,4.14.62,5.65,1.85,1.14.93,1.96,2.1,2.48,3.53.48-1.43,1.31-2.6,2.47-3.53,1.54-1.23,3.48-1.85,5.8-1.85,2.71,0,4.89.83,6.56,2.5,1.66,1.67,2.5,3.96,2.5,6.87v16.5ZM321.52,568.76c-.94,1.99-2.27,3.57-4.01,4.71-1.74,1.14-3.83,1.72-6.3,1.72-2.08,0-3.96-.48-5.62-1.43-1.66-.96-2.95-2.4-3.85-4.35l-.36,5.15h-3.12v-37.95h3.59v17.65c.87-1.87,2.12-3.27,3.77-4.19,1.65-.92,3.55-1.38,5.7-1.38,2.46,0,4.56.58,6.27,1.74,1.72,1.17,3.04,2.75,3.96,4.76.92,2.01,1.38,4.28,1.38,6.82s-.47,4.75-1.41,6.74ZM350.83,568.76c-1.11,1.99-2.63,3.57-4.56,4.71s-4.12,1.72-6.59,1.72-4.62-.57-6.56-1.72c-1.94-1.14-3.47-2.72-4.58-4.71-1.11-2-1.66-4.26-1.66-6.79s.55-4.89,1.66-6.87c1.11-1.98,2.64-3.54,4.58-4.68,1.94-1.14,4.13-1.72,6.56-1.72s4.66.57,6.59,1.72,3.44,2.71,4.56,4.68c1.11,1.98,1.66,4.27,1.66,6.87s-.56,4.8-1.66,6.79ZM378.89,574.56h-3.12l-.47-4.27c-.7,1.46-1.85,2.64-3.46,3.54-1.61.9-3.41,1.35-5.39,1.35-2.95,0-5.22-.94-6.79-2.81-1.58-1.87-2.37-4.34-2.37-7.39v-15.67h3.64v14.53c0,2.98.62,5.08,1.85,6.3,1.23,1.22,2.84,1.82,4.82,1.82,2.43,0,4.32-.79,5.68-2.37,1.35-1.58,2.03-3.86,2.03-6.84v-13.38h3.59v25.2ZM399.55,552.28h-1.98c-2.33,0-4.17.71-5.54,2.14-1.37,1.42-2.06,3.35-2.06,5.78v14.37h-3.59v-25.2h3.38l.25,3.95c.44-1.18,1.18-2.18,2.25-2.99,1.35-1.02,3-1.54,4.95-1.54.38,0,.76.02,1.15.05.38.04.78.11,1.2.21v3.23ZM426.78,573.78c0,2.5-.5,4.68-1.51,6.53-1.01,1.86-2.42,3.3-4.24,4.32-1.82,1.02-3.97,1.54-6.43,1.54-3.23,0-5.9-.87-8.02-2.6-2.12-1.74-3.44-4.11-3.96-7.13h3.59c.38,2.05,1.31,3.63,2.79,4.76,1.47,1.13,3.36,1.69,5.65,1.69,2.6,0,4.69-.8,6.25-2.4,1.56-1.59,2.34-3.71,2.34-6.35v-6.02c-.75,1.55-1.83,2.82-3.25,3.81-1.79,1.23-3.88,1.85-6.27,1.85s-4.46-.57-6.19-1.72c-1.74-1.15-3.06-2.67-3.98-4.56-.92-1.89-1.38-3.98-1.38-6.27s.47-4.38,1.41-6.27c.94-1.89,2.29-3.41,4.06-4.55s3.89-1.72,6.35-1.72,4.44.59,6.14,1.77c1.4.97,2.45,2.32,3.2,4l.34-5.09h3.12v24.42Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M419,553.09c-1.27-.81-2.84-1.22-4.71-1.22s-3.34.42-4.61,1.27c-1.27.85-2.24,1.98-2.91,3.38-.68,1.41-1.01,2.96-1.01,4.66s.34,3.31,1.01,4.74c.68,1.42,1.65,2.55,2.91,3.38,1.27.83,2.79,1.25,4.56,1.25s3.36-.41,4.66-1.22c1.3-.81,2.3-1.92,2.99-3.33.69-1.41,1.04-3.01,1.04-4.82s-.34-3.4-1.01-4.79-1.65-2.49-2.91-3.31Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+        <g>
+          <path
+            d="M245.29,260.42c-1.44-1.47-3.34-2.21-5.7-2.21-1.74,0-3.25.4-4.55,1.2-1.3.8-2.3,1.93-2.99,3.41-.48,1.02-.78,2.18-.93,3.46h16.34c0-2.43-.72-4.38-2.16-5.86Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M36.37,222.47v79.75h414.32v-79.75H36.37ZM221.79,281.01h-3.12l-.36-5.15c-.9,1.94-2.19,3.39-3.85,4.35-1.66.96-3.56,1.43-5.67,1.43-2.43,0-4.51-.57-6.25-1.72-1.74-1.14-3.07-2.72-4.01-4.71-.94-2-1.41-4.24-1.41-6.74s.47-4.81,1.41-6.82,2.28-3.6,4.03-4.76c1.75-1.16,3.88-1.74,6.38-1.74,2.08,0,3.93.45,5.54,1.35,1.61.9,2.85,2.27,3.72,4.11v-17.54h3.59v37.95ZM251.05,268.88h-20.02c.12,2.93.88,5.24,2.29,6.92,1.55,1.84,3.74,2.76,6.59,2.76,2.05,0,3.73-.43,5.05-1.3,1.32-.87,2.2-2.14,2.65-3.8h3.44c-.66,2.64-1.94,4.66-3.85,6.06-1.91,1.41-4.36,2.11-7.34,2.11-2.43,0-4.57-.56-6.43-1.67-1.86-1.11-3.31-2.65-4.35-4.63-1.04-1.98-1.56-4.28-1.56-6.92s.51-4.9,1.53-6.9c1.02-1.99,2.44-3.56,4.24-4.68,1.8-1.13,3.91-1.69,6.3-1.69s4.3.5,6.01,1.51,3.05,2.4,4.01,4.19c.95,1.79,1.43,3.88,1.43,6.27v1.77Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M214.14,259.67c-1.28-.87-2.86-1.3-4.74-1.3s-3.36.43-4.66,1.3c-1.3.87-2.29,2.05-2.97,3.54-.68,1.49-1.01,3.21-1.01,5.15s.34,3.67,1.01,5.18c.68,1.51,1.66,2.69,2.97,3.54,1.3.85,2.85,1.27,4.66,1.27s3.45-.42,4.74-1.27c1.28-.85,2.27-2.03,2.97-3.54.69-1.51,1.04-3.24,1.04-5.18s-.35-3.66-1.04-5.15c-.7-1.49-1.68-2.67-2.97-3.54Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+        <g>
+          <path
+            d="M15.58,322.16v79.75h476.31v-79.75H15.58ZM236.25,380.69h-3.59v-37.95h3.59v37.95ZM263.89,380.69h-3.02l-.47-4.43c-.76,1.6-1.97,2.84-3.62,3.72-1.65.89-3.53,1.33-5.65,1.33-2.67,0-4.76-.69-6.27-2.08-1.51-1.39-2.27-3.18-2.27-5.36,0-2.32.83-4.18,2.5-5.57s3.97-2.2,6.92-2.45l8.33-.68v-.84c0-1.66-.3-2.97-.89-3.93-.59-.96-1.39-1.62-2.4-2-1.01-.38-2.13-.57-3.38-.57-2.22,0-3.94.47-5.16,1.41-1.21.94-1.82,2.26-1.82,3.95h-3.28c0-1.7.44-3.17,1.3-4.42.87-1.25,2.08-2.22,3.64-2.92,1.56-.69,3.37-1.04,5.41-1.04s3.71.34,5.21,1.02c1.49.68,2.68,1.72,3.57,3.12.89,1.41,1.33,3.22,1.33,5.44v16.29Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M247.81,370.23c-1.04.97-1.56,2.15-1.56,3.54,0,1.53.52,2.68,1.56,3.46,1.04.78,2.43,1.17,4.17,1.17,2.67,0,4.74-.85,6.19-2.55,1.46-1.7,2.19-3.9,2.19-6.61v-1.35l-7.44.57c-2.36.21-4.06.8-5.1,1.77Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+        <g>
+          <path
+            d="M184.86,421.84v79.75h265.12v-79.75H184.86ZM234.75,475.9c-1.15,1.77-2.77,3.13-4.87,4.09-2.1.96-4.56,1.43-7.37,1.43s-5.01-.46-7.03-1.38c-2.01-.92-3.58-2.21-4.71-3.88-1.13-1.66-1.73-3.62-1.8-5.88h6.67c.03,1.6.67,2.86,1.9,3.78,1.23.92,2.91,1.38,5.02,1.38,2.22,0,3.98-.46,5.28-1.38,1.3-.92,1.95-2.14,1.95-3.67,0-1.32-.35-2.37-1.04-3.15-.7-.78-1.82-1.33-3.38-1.64l-6.04-1.35c-3.23-.69-5.67-1.95-7.34-3.77-1.66-1.82-2.5-4.21-2.5-7.16,0-2.22.57-4.18,1.72-5.88,1.15-1.7,2.73-3.03,4.76-3.98,2.03-.95,4.38-1.43,7.05-1.43s4.8.45,6.69,1.35c1.89.9,3.37,2.19,4.42,3.85,1.06,1.66,1.62,3.64,1.69,5.93h-6.66c-.07-1.63-.66-2.91-1.77-3.83-1.11-.92-2.6-1.38-4.48-1.38-2.05,0-3.69.46-4.92,1.38-1.23.92-1.85,2.18-1.85,3.78,0,1.35.37,2.42,1.12,3.2.74.78,1.92,1.36,3.51,1.75l5.99,1.35c3.26.7,5.69,1.88,7.29,3.57,1.6,1.68,2.39,3.95,2.39,6.79,0,2.33-.57,4.37-1.72,6.14ZM265.99,480.74h-6.35v-14.31c0-1.46-.23-2.66-.68-3.59-.45-.94-1.08-1.64-1.87-2.11-.8-.47-1.74-.7-2.81-.7-1.32,0-2.45.29-3.41.86-.96.57-1.68,1.36-2.19,2.37-.5,1.01-.76,2.15-.76,3.44v14.06h-6.35v-38.73h6.4v16.35c.8-1.25,1.92-2.25,3.36-3,1.44-.74,3.1-1.12,4.97-1.12,3.16,0,5.56.96,7.21,2.86,1.65,1.91,2.47,4.51,2.47,7.81v15.83ZM296.44,474.88c-1.18,2.03-2.8,3.62-4.87,4.76s-4.43,1.72-7.11,1.72-5-.57-7.08-1.72c-2.08-1.14-3.71-2.73-4.89-4.76-1.18-2.03-1.77-4.38-1.77-7.05s.59-5.03,1.77-7.06c1.18-2.03,2.81-3.62,4.89-4.76,2.08-1.14,4.44-1.72,7.08-1.72s5.04.57,7.11,1.72c2.06,1.15,3.69,2.73,4.87,4.76,1.18,2.03,1.77,4.38,1.77,7.06s-.59,5.02-1.77,7.05ZM325.54,480.74h-5.57l-.47-3.8c-.56,1.32-1.62,2.4-3.2,3.23-1.58.84-3.38,1.25-5.39,1.25-2.71,0-4.85-.74-6.43-2.21-1.58-1.47-2.37-3.41-2.37-5.8s.82-4.24,2.47-5.65c1.65-1.41,4-2.23,7.06-2.47l7.7-.62v-.57c0-1.18-.22-2.12-.65-2.84-.44-.71-1.05-1.23-1.85-1.56-.8-.33-1.74-.49-2.81-.49-1.87,0-3.31.37-4.32,1.12-1.01.75-1.51,1.8-1.51,3.15h-5.41c0-1.87.48-3.49,1.43-4.87.95-1.37,2.31-2.44,4.06-3.2,1.75-.76,3.77-1.14,6.06-1.14s4.32.41,5.99,1.22c1.66.82,2.95,2.03,3.85,3.64.9,1.62,1.35,3.62,1.35,6.01v15.62ZM356.2,480.74h-6.35v-14.31c0-1.46-.23-2.66-.68-3.59-.45-.94-1.08-1.64-1.87-2.11-.8-.47-1.74-.7-2.81-.7-1.32,0-2.45.29-3.41.86-.96.57-1.68,1.36-2.19,2.37s-.76,2.15-.76,3.44v14.06h-6.35v-38.73h6.4v16.35c.8-1.25,1.92-2.25,3.36-3,1.44-.74,3.1-1.12,4.97-1.12,3.16,0,5.56.96,7.21,2.86,1.65,1.91,2.47,4.51,2.47,7.81v15.83Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M309.82,470.62c-.83.64-1.25,1.48-1.25,2.52,0,1.14.38,2.01,1.15,2.6.76.59,1.84.88,3.23.88,1.98,0,3.55-.55,4.71-1.66,1.16-1.11,1.74-2.65,1.74-4.63v-1.35l-5.36.42c-1.98.17-3.38.58-4.22,1.22Z"
+            style={{
+              fill: color,
+            }}
+          />
+          <path
+            d="M288.32,460.96c-1.11-.66-2.39-.99-3.85-.99s-2.73.33-3.83.99c-1.09.66-1.96,1.57-2.6,2.73-.64,1.16-.96,2.54-.96,4.14s.32,2.93.96,4.11,1.51,2.1,2.6,2.76c1.09.66,2.37.99,3.83.99s2.74-.33,3.85-.99c1.11-.66,1.98-1.58,2.6-2.76.62-1.18.94-2.55.94-4.11s-.31-2.98-.94-4.14c-.62-1.16-1.49-2.07-2.6-2.73Z"
+            style={{
+              fill: color,
+            }}
+          />
+        </g>
+      </svg>
+    </div>
   )
 }
